Push the default strings.xml
Engrish is not my first language so it's bound to errors. Just pushing this in case anyone wants to contribute with language fixes to make the overal app experience better for everyone.

Please don't submit translations, I have no plans for localization it's just a pain in the arse to maintain with an app like this.

Also if there's weird ass symbols in the strings, they're likely arguments for getString() calls, so don't remove them.

Thanks!